+++ card_prisoner/__init__.py
@@ -1,0 +1,26 @@
+"""
+The project is designed under the MVC pattern. shown as below.
+
+model:
+* player.py
+
+view:
+* view.py
+* sidebar.py
+* item_list.py
+* textbox.py
+
+controller:
+* card_prisoner.py
+
+others:
+* constants.py
+* layout.py
+* messages.py
+* static_preview.py
+
+* item.py
+* talent.py
+"""
+
+from card_prisoner.card_prisoner import main